fix(async-combobox): Set clear button type as "button"
When having an invalid form and pressing enter, the input
value of the AsyncCombobox was cleared. This happened by
calling onClick handler for the clear button `handleClear`
of the AsyncCombobox.

But in HTML, any <button> inside a <form> defaults to
type="submit" unless specified otherwise. So when pressing
"Enter" while focused on an input field, it triggers the
first button with type="submit" in the form, or submits the
form itself — which called the clear button's onClick handler.