fix(node-dialogs): Clean-up node dialog state handling
Open / close is now only managed through useEffect, the
dialog is opened if pendingNode or currentEditNode is set
and closed if the parameter is set to null.

There was a bug in the CreateNodeDialog the the pendingNodeData
was not set to null on closing. This led to the behavior that the
dialog was closed, but on a reload it would be opened again.